+++ logging_manager/logging_manager/lib/src/default_logging_manager.dart
@@ -1,0 +1,8 @@
+import 'package:logging_manager/logging_manager.dart';
+import 'package:logging_manager/src/stack_frame.dart';
+
+/// The default logging manager that uses the a child from the root Logger.
+final loggingManager = LoggingManager(
+  logger: Logger('DefaultLogger'),
+  tree: kReleaseMode ? null : PrintingColoredLogsTree(),
+);