+++ logging_manager/logging_manager/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+- Update interface of logging manager.
+- Add a default logging manager getter `loggingManager`.
+
 ## 1.0.0+1
 
 - Documentation updates.